Introduce pipeline / component creation from DataCheckActions (#1907)
* init

* add test

* fix release notes
